[Unrated] Title: 더하기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Unrated/14918. 더하기/더하기.cc
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    long a, b;
+    int a, b;
     cin >> a >> b;
     cout << a+b;
 }
